+++ Demo/Keyboard/KeyboardViewController.swift
@@ -33,9 +33,22 @@ class KeyboardViewController: KeyboardInputViewController {
         /// ‼️ Set up this keyboard with `.keyboardKitDemo`.
         super.setup(for: .keyboardKitDemo)
 
-        /// 💡 Make basic state & service customizations.
-        setupDemoServices()
-        setupDemoState()
+        /// 💡 You can replace services with custom services.
+        services.autocompleteService = services.autocompleteService
+
+        /// 💡 Dock the keyboard view to any horizontal edge.
+        // state.keyboardContext.settings.keyboardDockEdge = .leading
+
+        /// 💡 Configure the space key's long press behavior.
+        state.keyboardContext.settings.spaceLongPressBehavior = .moveInputCursor
+
+        /// 💡 Configure the space key's additional actions.
+        // state.keyboardContext.settings.spaceContextMenuLeading = .locale
+        state.keyboardContext.settings.spaceContextMenuTrailing = .locale
+
+        /// 💡 Customize keyboard feedback.
+        // state.feedbackContext.settings.isAudioFeedbackEnabled = false
+        // state.feedbackContext.settings.isHapticFeedbackEnabled = false
     }
     
     /// This function is called when the controller needs to
@@ -43,13 +56,11 @@ class KeyboardViewController: KeyboardInputViewController {
     ///
     /// Call `setupKeyboardView(_:)` here to set up a custom
     /// keyboard view or customize the default `KeyboardView`.
+    ///
+    /// Don't call `super.viewWillSetupKeyboardView()` since
+    /// that will cause two layout calculations.
     override func viewWillSetupKeyboardView() {
-        /// 💡 Don't call `super.viewWillSetupKeyboardView()`.
-        // super.viewWillSetupKeyboardView()
 
-        /// 💡 Call `setupKeyboardView(...)` to customize or
-        /// replace the standard `KeyboardView`.
-        ///
         /// Return `$0.view` to return the standard view, or
         /// return a custom view for the provided parameters.
         setupKeyboardView { controller in
@@ -71,39 +82,25 @@ class KeyboardViewController: KeyboardInputViewController {
                 }
             )
             // .autocorrectionDisabled()
+            // 💡 Setup custom callout actions for the k key.
+            .keyboardCalloutActions { params in
+                let action = params.action
+                let context = controller.state.keyboardContext
+                let custom = String("keyboardkit".reversed())
+                if action.shouldUseCustomCallouts { return .init(characters: custom) }
+                return params.standardCalloutActions(for: context)
+            }
             // .keyboardToolbarStyle(.init(backgroundColor: .red))
         }
     }
 }
 
-private extension KeyboardViewController {
+private extension KeyboardAction {
 
-    /// Make demo-specific changes to your keyboard services.
-    func setupDemoServices() {
-
-        /// 💡 You can replace any service with a custom service.
-        services.autocompleteService = services.autocompleteService
-    }
-
-    /// Make demo-specific changes to your keyboard's state.
-    func setupDemoState() {
-        
-        /// 💡 This enable more locales.
-        state.keyboardContext.locales = [.english, .spanish]
-        
-        /// 💡 This overrides the standard enabled locales.
-        state.keyboardContext.settings.addedLocales = [.init(.english), .init(.swedish)]
-        
-        /// 💡 Dock the keyboard to any horizontal edge.
-        // state.keyboardContext.settings.keyboardDockEdge = .leading
-
-        /// 💡 Configure the space key's long press behavior and trailing action.
-        state.keyboardContext.settings.spaceLongPressBehavior = .moveInputCursor
-        // state.keyboardContext.settings.spaceContextMenuLeading = .locale
-        state.keyboardContext.settings.spaceContextMenuTrailing = .locale
-
-        /// 💡 Customize keyboard feedback.
-        // state.feedbackContext.settings.isAudioFeedbackEnabled = false
-        // state.feedbackContext.settings.isHapticFeedbackEnabled = false
+    var shouldUseCustomCallouts: Bool {
+        switch self {
+        case .character(let char): char.lowercased() == "k"
+        default: false
+        }
     }
 }